[FAB-17510] Change AddOrgToConsortium() to modify the provided config
