+++ pkg/config/consortiums.go
@@ -7,6 +7,7 @@ SPDX-License-Identifier: Apache-2.0
 package config
 
 import (
+	"errors"
 	"fmt"
 
 	"github.com/golang/protobuf/proto"
@@ -19,6 +20,39 @@ import (
 type Consortium struct {
 	Name          string
 	Organizations []*Organization
+}
+
+// AddOrgToConsortium adds an org definition to a named consortium in a given
+// channel configuration.
+func AddOrgToConsortium(config *cb.Config, org *Organization, consortium string, mspConfig *mb.MSPConfig) error {
+	if org == nil {
+		return errors.New("organization is required")
+	}
+	if consortium == "" {
+		return errors.New("consortium is required")
+	}
+
+	consortiumsGroup, ok := config.ChannelGroup.Groups[ConsortiumsGroupKey]
+	if !ok {
+		return errors.New("consortiums group does not exist")
+	}
+
+	consortiumGroup, ok := consortiumsGroup.Groups[consortium]
+	if !ok {
+		return fmt.Errorf("consortium '%s' does not exist", consortium)
+	}
+
+	orgGroup, err := newConsortiumOrgGroup(org, mspConfig)
+	if err != nil {
+		return fmt.Errorf("failed to create consortium org: %v", err)
+	}
+
+	if consortiumGroup.Groups == nil {
+		consortiumGroup.Groups = map[string]*cb.ConfigGroup{}
+	}
+	consortiumGroup.Groups[org.Name] = orgGroup
+
+	return nil
 }
 
 // NewConsortiumsGroup returns the consortiums component of the channel configuration. This element is only defined for
