+++ pkg/config/consortiums_test.go
@@ -7,13 +7,232 @@ SPDX-License-Identifier: Apache-2.0
 package config
 
 import (
+	"bytes"
 	"testing"
 
+	"github.com/golang/protobuf/proto"
 	cb "github.com/hyperledger/fabric-protos-go/common"
 	mb "github.com/hyperledger/fabric-protos-go/msp"
+	"github.com/hyperledger/fabric/common/tools/protolator"
 	. "github.com/onsi/gomega"
 )
 
+func TestAddOrgToConsortium(t *testing.T) {
+	gt := NewGomegaWithT(t)
+
+	config := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Consortiums": {
+					Groups: map[string]*cb.ConfigGroup{
+						"test-consortium": {},
+					},
+				},
+			},
+		},
+	}
+
+	org := &Organization{
+		Name:     "Org1",
+		ID:       "Org1MSP",
+		Policies: applicationOrgStandardPolicies(),
+	}
+
+	err := AddOrgToConsortium(config, org, "test-consortium", &mb.MSPConfig{})
+	gt.Expect(err).NotTo(HaveOccurred())
+
+	expectedConfig := `
+{
+	"channel_group": {
+		"groups": {
+			"Consortiums": {
+				"groups": {
+					"test-consortium": {
+						"groups": {
+							"Org1": {
+								"groups": {},
+								"mod_policy": "Admins",
+								"policies": {
+									"Admins": {
+										"mod_policy": "Admins",
+										"policy": {
+											"type": 3,
+											"value": {
+												"rule": "MAJORITY",
+												"sub_policy": "Admins"
+											}
+										},
+										"version": "0"
+									},
+									"Endorsement": {
+										"mod_policy": "Admins",
+										"policy": {
+											"type": 3,
+											"value": {
+												"rule": "MAJORITY",
+												"sub_policy": "Endorsement"
+											}
+										},
+										"version": "0"
+									},
+									"LifecycleEndorsement": {
+										"mod_policy": "Admins",
+										"policy": {
+											"type": 3,
+											"value": {
+												"rule": "MAJORITY",
+												"sub_policy": "Endorsement"
+											}
+										},
+										"version": "0"
+									},
+									"Readers": {
+										"mod_policy": "Admins",
+										"policy": {
+											"type": 3,
+											"value": {
+												"rule": "ANY",
+												"sub_policy": "Readers"
+											}
+										},
+										"version": "0"
+									},
+									"Writers": {
+										"mod_policy": "Admins",
+										"policy": {
+											"type": 3,
+											"value": {
+												"rule": "ANY",
+												"sub_policy": "Writers"
+											}
+										},
+										"version": "0"
+									}
+								},
+								"values": {
+									"MSP": {
+										"mod_policy": "Admins",
+										"value": {
+											"config": null,
+											"type": 0
+										},
+										"version": "0"
+									}
+								},
+								"version": "0"
+							}
+						},
+						"mod_policy": "",
+						"policies": {},
+						"values": {},
+						"version": "0"
+					}
+				},
+				"mod_policy": "",
+				"policies": {},
+				"values": {},
+				"version": "0"
+			}
+		},
+		"mod_policy": "",
+		"policies": {},
+		"values": {},
+		"version": "0"
+	},
+	"sequence": "0"
+}
+`
+
+	expectedConfigProto := cb.Config{}
+	err = protolator.DeepUnmarshalJSON(bytes.NewBufferString(expectedConfig), &expectedConfigProto)
+	gt.Expect(err).To(BeNil())
+	gt.Expect(proto.Equal(config, &expectedConfigProto)).To(BeTrue())
+}
+
+func TestAddOrgToConsortiumFailures(t *testing.T) {
+	t.Parallel()
+
+	baseConfig := &cb.Config{
+		ChannelGroup: &cb.ConfigGroup{
+			Groups: map[string]*cb.ConfigGroup{
+				"Consortiums": {
+					Groups: map[string]*cb.ConfigGroup{
+						"test-consortium": {},
+					},
+				},
+			},
+		},
+	}
+
+	org := &Organization{
+		Name:     "test-org",
+		ID:       "test-org-msp-id",
+		Policies: applicationOrgStandardPolicies(),
+	}
+
+	for _, test := range []struct {
+		name        string
+		org         *Organization
+		consortium  string
+		config      *cb.Config
+		expectedErr string
+	}{
+		{
+			name:        "When the organization is nil",
+			org:         nil,
+			consortium:  "test-consortium",
+			config:      baseConfig,
+			expectedErr: "organization is required",
+		},
+		{
+			name:        "When the consortium name is not specified",
+			org:         org,
+			consortium:  "",
+			config:      baseConfig,
+			expectedErr: "consortium is required",
+		},
+		{
+			name:       "When the config doesn't contain a consortiums group",
+			org:        org,
+			consortium: "test-consortium",
+			config: &cb.Config{
+				ChannelGroup: &cb.ConfigGroup{
+					Groups: map[string]*cb.ConfigGroup{},
+				},
+			},
+			expectedErr: "consortiums group does not exist",
+		},
+		{
+			name:        "When the config doesn't contain the consortium",
+			org:         org,
+			consortium:  "what-the-what",
+			config:      baseConfig,
+			expectedErr: "consortium 'what-the-what' does not exist",
+		},
+		{
+			name: "When the config doesn't contain the consortium",
+			org: &Organization{
+				Name: "test-msp",
+				ID:   "test-org-msp-id",
+				Policies: map[string]*Policy{
+					"Admins": nil,
+				},
+			},
+			consortium:  "test-consortium",
+			config:      baseConfig,
+			expectedErr: "failed to create consortium org: no Admins policy defined",
+		},
+	} {
+		test := test
+		t.Run(test.name, func(t *testing.T) {
+			t.Parallel()
+			gt := NewGomegaWithT(t)
+
+			err := AddOrgToConsortium(test.config, test.org, test.consortium, &mb.MSPConfig{})
+			gt.Expect(err).To(MatchError(test.expectedErr))
+		})
+	}
+}
 func TestNewConsortiumsGroup(t *testing.T) {
 	t.Parallel()
